Add basic client response
This isn't an ideal solution but it works for now. As part of the log
entry we store a RequestId which is a unique ID identifying the request
from the client. When an entry gets applied to the state machine, the
node sends a ClientResponse message containing that entry.

In the real world it's only the node that received the client request
that can send this message, because the other nodes wouldn't know who to
send it to. In the current setup, all nodes send this request for all
committed entries, so there's a load of duplicate messages being sent.