Retry client requests when leader is discovered
When we make a request to a server which isn't a leader, but does know
who the leader is, we can retry the request to the leader.

Whether a server knows the leader depends on them having voted for them
in the past, so it isn't foolproof. If a server goes down and comes back
up, it may rejoin a stable cluster and therefore not have voted for
anyone. We should consider storing the ID of the server that sent the
last successful AppendEntries request and using that instead.